move postAnyAddress to shelley and use proper network discrimination there plus code reshuffling
